Fordele Authorization (én plugin) i 'vefifiser innlogging', 'organisasjoner en bruker har tilgang til i Altinn' og 'organisasjoner en bruker har enkeltrettigheter til' (flere plugins)
