+++ src/main/kotlin/no/nav/pia/sykefravarsstatistikk/api/OrganisasjonerRoutes.kt
@@ -5,18 +5,18 @@ import io.ktor.server.response.respond
 import io.ktor.server.routing.Route
 import io.ktor.server.routing.get
 import io.ktor.server.routing.route
-import no.nav.pia.sykefravarsstatistikk.api.auth.TilgangerKey
+import no.nav.pia.sykefravarsstatistikk.api.auth.AltinnTilgangerKey
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 
 fun Route.organisasjoner() {
     val logger: Logger = LoggerFactory.getLogger(this::class.java)
 
-    route("/sykefravarsstatistikk/{orgnr}/organisasjoner") {
+    route("/sykefravarsstatistikk/organisasjoner") {
         route("/tilgang") {
             get {
                 logger.info("Henter organisasjoner")
-                val tilganger = call.attributes[TilgangerKey]
+                val tilganger = call.attributes[AltinnTilgangerKey]
 
                 call.respond(
                     status = HttpStatusCode.OK,
@@ -28,7 +28,7 @@ fun Route.organisasjoner() {
         route("/enkeltrettighet") {
             get {
                 logger.info("Henter organisasjoner med enkeltrettighet")
-                val tilganger = call.attributes[TilgangerKey]
+                val tilganger = call.attributes[AltinnTilgangerKey]
 
                 call.respond(
                     status = HttpStatusCode.OK,
