+++ src/main/kotlin/no/nav/pia/sykefravarsstatistikk/api/SykefraværsstatistikkRoutes.kt
@@ -7,7 +7,7 @@ import io.ktor.server.routing.get
 import io.ktor.server.routing.route
 import no.nav.pia.sykefravarsstatistikk.api.aggregering.AggregertStatistikkService
 import no.nav.pia.sykefravarsstatistikk.api.auth.OverordnetEnhetKey
-import no.nav.pia.sykefravarsstatistikk.api.auth.TilgangerKey
+import no.nav.pia.sykefravarsstatistikk.api.auth.VerifiserteTilgangerKey
 import no.nav.pia.sykefravarsstatistikk.api.auth.UnderenhetKey
 import no.nav.pia.sykefravarsstatistikk.persistering.KvartalsvisSykefraværshistorikkService
 import org.slf4j.Logger
@@ -26,7 +26,7 @@ fun Route.sykefraværsstatistikk(
                 kvartalsvisSykefraværshistorikkService.hentSykefraværshistorikk(
                     overordnetEnhet = call.attributes[OverordnetEnhetKey],
                     underenhet = call.attributes[UnderenhetKey],
-                    tilganger = call.attributes[TilgangerKey],
+                    tilganger = call.attributes[VerifiserteTilgangerKey],
                 ).map {
                     call.respond(
                         status = HttpStatusCode.OK,
@@ -47,7 +47,7 @@ fun Route.sykefraværsstatistikk(
 
                 aggregertStatistikkService.hentAggregertStatistikk(
                     underenhet = call.attributes[UnderenhetKey],
-                    tilganger = call.attributes[TilgangerKey],
+                    tilganger = call.attributes[VerifiserteTilgangerKey],
                 ).map {
                     call.respond(
                         status = HttpStatusCode.OK,
