+++ src/test/kotlin/no/nav/pia/sykefravarsstatistikk/helper/TestContainerHelper.kt
@@ -156,21 +156,15 @@ class TestContainerHelper {
                 config = config,
             ) ?: fail("Feil ved henting av publiseringsdato, mottok ikke respons")
 
-        suspend fun hentOrganisasjonerTilgangResponse(
-            orgnr: String,
-            config: HttpRequestBuilder.() -> Unit = {},
-        ): HttpResponse =
+        suspend fun hentOrganisasjonerTilgangResponse(config: HttpRequestBuilder.() -> Unit = {}): HttpResponse =
             applikasjon.performGet(
-                url = "/sykefravarsstatistikk/$orgnr/organisasjoner/tilgang",
+                url = "/sykefravarsstatistikk/organisasjoner/tilgang",
                 config = config,
             ) ?: fail("Feil ved henting av organisasjoner vedkommenede har tilgang til, mottok ikke respons")
 
-        suspend fun hentOrganisasjonerMedEnkeltrettighetResponse(
-            orgnr: String,
-            config: HttpRequestBuilder.() -> Unit = {},
-        ): HttpResponse =
+        suspend fun hentOrganisasjonerMedEnkeltrettighetResponse(config: HttpRequestBuilder.() -> Unit = {}): HttpResponse =
             applikasjon.performGet(
-                url = "/sykefravarsstatistikk/$orgnr/organisasjoner/enkeltrettighet",
+                url = "/sykefravarsstatistikk/organisasjoner/enkeltrettighet",
                 config = config,
             ) ?: fail("Feil ved henting av organisasjoner på enkelrettighet, mottok ikke respons")
